Add a quick script for generating test users
Add a (hacky) script for generating test users in cognito. The script
generates 3 users (admin, publisher and viewer) and a new team with the
viewer user as a member. These accounts use fake emails so they can not
be used for testing the new user invite and password reset
functionality, but they should allow a good range of testing with
different permissions.

This script is not intended to be robust. It does not have any error
handling, has minimal input validation, doesn't handle names with
special characters or spaces, has minimal documentation at best and
doesn't provide any mechanism to cleanup the generated users and teams
(manual cleanup is a pain). While something is better than nothing, some
further work in this area would be beneficial.