Remove JWT authentication from password change endpoint
The password change endpoint uses either a single use password reset
token or a single use session token in the body to authenticate the
request rather than the standard JWT token header. As such the extra
auth middleware was resulting in a 403 response and blocking the ability
to set the password.